Fixed opencl ell product wrong type bug

+++ honei/backends/opencl/product.cl
@@ -20,8 +20,8 @@
 
 #pragma OPENCL EXTENSION cl_khr_fp64 : enable
 
-__kernel void product_smell_dv_float(__global float * x, __global float * y, __global unsigned long * Aj, __global float * Ax,
-                                              __global unsigned long * Arl, unsigned long num_rows, unsigned long stride)
+__kernel void product_smell_dv_float(__global float * x, __global float * y, __global unsigned int * Aj, __global float * Ax,
+                                              __global unsigned int * Arl, unsigned int num_rows, unsigned int stride)
 {
     uint row = get_global_id(0);
 
@@ -31,13 +31,13 @@ __kernel void product_smell_dv_float(__global float * x, __global float * y, __g
     Aj += row;
     Ax += row;
 
-    const unsigned long max = Arl[row];
-    for(unsigned long n = 0; n < max ; n++){
+    const unsigned int max = Arl[row];
+    for(unsigned int n = 0; n < max ; n++){
         const float A_ij = *Ax;
 
         //if (A_ij != 0)
         {
-            const unsigned long col = *Aj;
+            const unsigned int col = *Aj;
             sum += A_ij * x[col];
         }
 
@@ -48,8 +48,8 @@ __kernel void product_smell_dv_float(__global float * x, __global float * y, __g
     y[row] = sum;
 }
 
-__kernel void product_smell_dv_double(__global double * x, __global double * y, __global unsigned long * Aj, __global double * Ax,
-                                               __global unsigned long * Arl, unsigned long num_rows, unsigned long stride)
+__kernel void product_smell_dv_double(__global double * x, __global double * y, __global unsigned int * Aj, __global double * Ax,
+                                               __global unsigned int * Arl, unsigned int num_rows, unsigned int stride)
 {
     uint row = get_global_id(0);
 
@@ -59,13 +59,13 @@ __kernel void product_smell_dv_double(__global double * x, __global double * y, 
     Aj += row;
     Ax += row;
 
-    const unsigned long max = Arl[row];
-    for(unsigned long n = 0; n < max ; n++){
+    const unsigned int max = Arl[row];
+    for(unsigned int n = 0; n < max ; n++){
         const double A_ij = *Ax;
 
         //if (A_ij != 0)
         {
-            const unsigned long col = *Aj;
+            const unsigned int col = *Aj;
             sum += A_ij * x[col];
         }
